+++ R/roxygenize.R
@@ -30,7 +30,7 @@ roxygenize <- function(package.dir = ".",
                        load_code = NULL,
                        clean = FALSE) {
 
-  base_path <- normalizePath(package.dir)
+  base_path <- normalizePath(package.dir, mustWork = TRUE)
   is_first <- roxygen_setup(base_path)
 
   encoding <- desc::desc_get("Encoding", file = base_path)[[1]]
@@ -53,6 +53,12 @@ roxygenize <- function(package.dir = ".",
 
   roclets <- lapply(roclets, roclet_find)
 
+  # Now load code
+  load_code <- find_load_strategy(load_code)
+  env <- load_code(base_path)
+  roxy_meta_set("env", env)
+  on.exit(roxy_meta_set("env", NULL), add = TRUE)
+
   # Tokenise each file
   blocks <- parse_package(base_path, env = NULL)
 
@@ -65,9 +71,6 @@ roxygenize <- function(package.dir = ".",
     base_path = base_path
   )
 
-  # Now load code
-  load_code <- find_load_strategy(load_code)
-  env <- load_code(base_path)
   blocks <- lapply(blocks, block_set_env, env = env)
 
   results <- lapply(roclets, roclet_process,