+++ tests/testthat/testNamespace/R/a.r
@@ -1,0 +1,6 @@
+#' @export
+f <- function(x) x
+
+
+#' @evalNamespace "export(g)"
+g <- function() 1