+++ NEWS.md
@@ -1,3 +1,5 @@
+# roxygen2 (development version)
+
 # roxygen2 7.0.2
 
 * `\example{}` escaping has been improved (again!) so that special escapes