+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # roxygen2 7.0.2
 
+* The warning for section titles spanning multiple lines now includes the hint "Did you forget a colon (:) at the end of the title?". (#994, @maelle)
+
 * `\example{}` escaping has been improved (again!) so that special escapes 
   within strings are correctly escaped (#990).