+++ NEWS.md
@@ -8,6 +8,9 @@
   package level documentation, where it can be used to suppress the
   auto-generated Description section (#1008).
 
+* roxygen2 now adds hyperlinks to R6 methods in the PDF manual as well,
+  (#1006).
+
 # roxygen2 7.0.2
 
 * `\example{}` escaping has been improved (again!) so that special escapes