+++ NEWS.md
@@ -1,5 +1,27 @@
 # roxygen2 (development version)
 
+* Can now document objects created with `delayedAssign()` by forcing
+  evaluation at documentation time (#1041)
+
+* `evalNamespace()` works again (#1022).
+
+* roxygen2 now keeps using Windows (CR LF) line endings for files that
+  already have CR LF line endings, and uses LF for new files (#989).
+
+* roxygen2 now does not fail for `@description NULL` and `@details NULL`.
+  Instead, these tags are ignored, except for `@description NULL` in
+  package level documentation, where it can be used to suppress the
+  auto-generated Description section (#1008).
+
+* roxygen2 now adds hyperlinks to R6 methods in the PDF manual as well,
+  (#1006).
+
+* Multiple `@format` tags are now combined (#1015).
+
+* roxygen2 now omits empty annotations (alternate text) for figures added
+  via markdown. This caused issues when generating pkgdown web sites
+  (#1051).
+
 # roxygen2 7.0.2
 
 * `\example{}` escaping has been improved (again!) so that special escapes 
@@ -193,7 +215,7 @@ You can override the default either by calling (e.g.) `roxygenise(load_code = "s
   released 2014-05-02, over 5 years ago.
 
 * Using the old `wrap` option will now trigger a warning, as hasn't worked
-  for quite some time. Supress the error by deleting the option from your
+  for quite some time. Suppress the error by deleting the option from your
   `DESCRIPTION`.
 
 ### Extending roxygen2