+++ NEWS.md
@@ -1,4 +1,4 @@
-# roxygen2 (development version)
+# roxygen2 7.1.0
 
 ## New features